+++ packages/web/index.jsx
@@ -7,7 +7,7 @@ import './polyfill';
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import { ConnectedRouter } from 'react-router-redux'
+import { ConnectedRouter } from 'react-router-redux';
 import { history, store } from '@qasir/store';
 
 import Application from './component/application.jsx';