+++ webpack.config.js
@@ -51,19 +51,19 @@ let configurationWebpack = {
   },
   module: {
       rules: [
-        {
-          enforce: 'pre',
-          test: /\.(jsx|js)$/,
-          include: /packages/,
-          exclude: /node_modules/,
-          loader: 'eslint-loader',
-          options: {
-            configFile: './.eslintrc',
-            emitWarning: true,
-            failOnError: false,
-            failOnWarning: false
-          }
-        },
+        // {
+        //   enforce: 'pre',
+        //   test: /\.(jsx|js)$/,
+        //   include: /packages/,
+        //   exclude: /node_modules/,
+        //   loader: 'eslint-loader',
+        //   options: {
+        //     configFile: './.eslintrc',
+        //     emitWarning: true,
+        //     failOnError: false,
+        //     failOnWarning: false
+        //   }
+        // },
         {
           test: /\.(jsx|js)?$/,
           exclude: /node_modules/,